sqlexec-webapp/requirements.txt
absl-py==1.0.0
anyio==3.4.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.4.1
asttokens==2.0.5
astunparse==1.6.3
attrs==21.4.0
autopep8==1.6.0
backcall==0.2.0
backports.entry-points-selectable==1.1.0
black==21.12b0
bleach==4.1.0
cachetools==5.0.0
certifi==2021.5.30
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.4
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.2
dj-database-url==0.5.0
Django==3.2.8
django-appconf==1.0.5
django-crispy-forms==1.13.0
django-heroku==0.3.1
django-responsive2==0.1.3
djangorestframework==3.13.1
entrypoints==0.3
executing==0.8.2
fastapi==0.70.0
filelock==3.0.12
flatbuffers==2.0
fonttools==4.29.0
gast==0.4.0
google-auth==2.4.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.43.0
gunicorn==20.1.0
h11==0.12.0
h5py==3.6.0
httptools==0.2.0
idna==3.3
importlib-metadata==4.10.1
ipykernel==6.7.0
ipython==8.0.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.1
Jinja2==3.0.3
jsonschema==4.4.0
jupyter==1.0.0
jupyter-client==7.1.2
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
libclang==12.0.0
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
mypy-extensions==0.4.3
mysql-connector-python==8.0.28
mysqlclient==2.1.0
nbclient==0.5.10
nbconvert==6.4.0
nbformat==5.1.3
nest-asyncio==1.5.4
notebook==6.4.7
numpy==1.22.1
oauthlib==3.1.1
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.0
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.9.0
pickleshare==0.7.5
Pillow==9.0.0
pipenv==2021.5.29
platformdirs==2.2.0
prometheus-client==0.12.0
prompt-toolkit==3.0.24
protobuf==3.19.3
psycopg2==2.9.3
psycopg2-binary==2.9.3
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
pycryptodome==3.12.0
pydantic==1.8.2
Pygments==2.11.2
pyparsing==3.0.7
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
pywin32==303
pywinpty==1.1.6
PyYAML==6.0
pyzmq==22.3.0
qtconsole==5.2.2
QtPy==2.0.0
requests==2.27.1
requests-oauthlib==1.3.0
rsa==4.8
scipy==1.7.3
seaborn==0.11.2
Send2Trash==1.8.0
six==1.16.0
sniffio==1.2.0
sqlparse==0.4.2
stack-data==0.1.4
starlette==0.16.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.23.1
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
toml==0.10.2
tomli==1.2.3
tornado==6.1
traitlets==5.1.1
typing_extensions==4.0.1
urllib3==1.26.8
uvicorn==0.15.0
virtualenv==20.7.0
virtualenv-clone==0.5.6
watchgod==0.7
wcwidth==0.2.5
webencodings==0.5.1
websockets==10.1
Werkzeug==2.0.2
whitenoise==5.3.0
widgetsnbextension==3.5.2
wrapt==1.13.3
zipp==3.7.0
